:recycle: refactor device editor
:lipstick: add visual effect to refetch com ports button
