+++ src/renderer/components/_features/[workspace]/editor/device/components/index.ts
@@ -1,0 +1,2 @@
+export * from './configuration-editor'
+export * from './pin-mapping-editor'
