+++ src/renderer/components/_features/[workspace]/editor/device/components/pin-mapping-editor.tsx
@@ -1,0 +1,5 @@
+const PinMappingEditor = () => {
+  return <div>Pin Mapping Editor</div>
+}
+
+export { PinMappingEditor }
